refactor(metadata): 归拢 BKBase 任务逻辑 --story=137192929
